VM should fail external calls returning uninitialized ExternalStructures
When an ExternalStructure is first instantiated in the image, its
structure template information and byte-size will be lazily-initialized
in the image. If the first time an ExternalStructure is instantiated is
by the VM in order to return a value from an external library call, then
the structure may remain uninitialized. In the case of return by value,
this means the VM will fail to allocate sufficient return buffer on the
stack and the call will crash (undesirable). In the case of return
by reference, the call won't crash, but the structure class will remain
uninitialized and so typically attempts to access the fields of the
structure will yield the wrong data. This typically results in knock-on
issues that are difficult to diagnose.
With this commit the VM checks that the return structure has a
non-zero byte size, and if so it will fail the call with a specific
failure code that the image call failure handling code can use to provide
a specific error report for the developer. When this occurs, the developer
needs to ensure that the structure is lazily initialized by sending the
relevant class a #ensureDefined message before attempting to make the
call that creates an instance on return.